fix: GCC 15 -Werror build (cstring, double-promotion, void* memcpy)
Three -Werror failures hit on GCC 15 (-O3 -Wall -Werror) that the
MSVC + earlier-GCC build paths didn't surface:

- ham.cpp: add <cstring> for std::memcpy, and route the SLF-stall
  hack from commit 3c68109 through `static_cast<void*>` to silence
  -Wclass-memaccess. SRGBColor has default member initializers (`{}`)
  which makes GCC treat it as non-trivial; the type is in fact
  trivially copyable and the 32-bit single-mov is the perf hack.

- quantize.cpp: cast color_space::WEIGHT_{L,A,B} to double in two
  k-means accumulator sites — dividing a `double` by a `float`
  WEIGHT triggers -Werror=double-promotion. Also explicit-cast the
  three OKLab subtractions in pnn_quantize seed-scoring so the
  difference computes in double directly instead of float→double.

- ssimulacra2.cpp: the inner Gaussian interior loop's
  `row[x + i - kBlurHalf]` mixed std::size_t (x), int (i), and
  int (kBlurHalf), tripping -Werror=sign-conversion. Cast i and
  kBlurHalf explicitly so the index is computed entirely in
  std::size_t (always non-negative inside the interior strip).

No code-gen change intended — these are diagnostic-only fixes.

+++ src/ssimulacra2.cpp
@@ -166,7 +166,8 @@ void gaussian_blur(const std::vector<float>& in,
             float s = 0.0f;
             for (int i = 0; i < kBlurSize; ++i) {
                 s += k[static_cast<std::size_t>(i)] *
-                     row[x + i - kBlurHalf];
+                     row[x + static_cast<std::size_t>(i)
+                           - static_cast<std::size_t>(kBlurHalf)];
             }
             trow[x] = s;
         }
